(core) link_image extended with anchor points
Previously, link image has always assumed the
upper left corner as the anchor point for
hierarchies, which severly limited the usefulness
of the feature for UI purposes (any and all
resizes to a parent meant manually recalculating
offsets). With this change, it is now possible
to specify which of the 5 (ul, ur, ll, lr, center)
anchor points one would like to define as the new
origo for the child.