(core+test) continued refactoring of GL use
PBOs and readbacks have been moved into
the agp part of the platform now, though
GLES3+ support is broken. We expose a few
new constants regarding shader language and
underlying graphics along with related regression
tests for the refactor. The big thing missing
still in the graphics refactor is FBO use with
the few minor things remaining are blend-
state and 3dbase use.