(core) streaming upload fixes
Moving streaming uploads to a separate platform
layer brought on the expected amount of painful
problems. These should be fixed now, but before
proceeding we should:

a. extend the benchmarking / regression suite
   with a frameserver that counts upwards each
   video update.

b. write a benchmark suite for PBO transfers
   that compares no-alpha-copy flag against
   regular ones.

c. write a regression test that checks integrity
   and sequence of frames where multiple objects
   share the same frameserver (should still only
   update once per frame), along with a  calctgt
   that makes sure the sequence steps are right.

After that, most this refactor needs is to move
out of 3dbase, create GLES2/3 compatible agp
versions and get rid of the GLEW req. for GL21.