(core+) shader management moved to platform
Moved the shader management to the agp/video platform where
it belongs. Completed dropping GLEW as a dependency and
now managed extension etc. mapping internally.

The EGL-DRI platform now works correctly with GL21
(including the rendertarget / readback parts), though
still not 100% possible to compile-time switch between
EGL3+ and GL21 due to gles3.c not being updated yet.

We are still in a broken state with the libretro frameserver
that should now start using the agp layer rather than its
own GL routines, then we are closing in on merging back to master.