(support) mouse script fixes, composition surface
Added first draft of composition surface (fairly
incomplete still) which is intended to make the
task of setting up / managing creation and decoration
of subsurfaces for both window manager purposes
and for specialized composition targets (e.g.
remoting and recording).

Fixed how mouse scripts managed absolute inputs,
and added a first test interactive script that uses
the composition surface but for the purpose of
developing / testing multi-display support.

While the multi-display support is intended for
lower level management (and a large reason for the
GL etc. refactor), the development is rather tedious
and crash-prone so to ease things and make sure test-cases
work properly, we use arcan_lwa in combination with
target_alloc for this purpose.

Every time arcan_lwa receives a new subsegment on the
primary shmif segment, we treat that as a "hotplug"
event of a secondary/tertiary etc. screen.