(platform) egl-dri segfault workarounds
The KMS/EGL/GBM combination is rather crash-
prone even for situations we can't reasonably
test. To make this less of an issue, we temporarily
install a SIGSEG handler that at least outputs
which step it is that would've crashed (typically
GBM trying to create a SCANOUT buffer when we
already have a drmMaster through X).