(core) added coverage tracing to Lua interface
defining LUA_TRACE_COVERAGE when compiling arcan_lua
will enable the tracing Lua->engine calls along with
the basic type of the contents of the stack into an
arcan.coverage file that will be saved in the SYS_DEBUG
namespace.

This is a CSV formated (;) file,
timestamp(ms);symbol_name;arg_1_type;arg_2_type

The intended use is to run the basic test-suites,
joining and storing unique entries in the arcan.coverage
file.

In the end, we can compare with the parsing done with
docgen to estimate (shallow) the rate of coverage compared
to the set of available functions.

FossilOrigin-Name: dea60d5d84fb60714c9217da6f16c937026c382e4f028085c8ace06540aa897a