(platform) build fixes for egl-gles
GLES- specific platform for primarily starving ARM devices
has been brought up to speed in regards to changed in the
video_platform. Unfortunately, the extension / symbol mapping
do not work correctly for a number of devices e.g. RPI.

Missing is to go through the list of extensions we want to
map, and only define those that are actual extensions.
(Now we map/track all needed functions which isn't particularly
EGL/etc. conformant but working on a lot of devices).