(core) added flag state for switching resolve
One of the more ill-designed and performance
degrading problems at the moment is that all
objects are treated as 3D (quaternions for
orientation, modelview matrices etc.) which
is incredibly costly for the most common case
i.e. 2D rendering. This is a legacy issue
that we are on the first steps of resolving.