(event) added record/replay/panic button
Build fixes for X11 platform caused by FreeBSD fixes.
Added support for additional environment variables in the
event layer:
ARCAN_EVENT_RECORD=file (record input events to file)
ARCAN_EVENT_REPLAY=file (replay input events from file)

This feature is planned to help with testing, fault injection
and benchmarking. The drawback is that timing is currently aligned
to frame-by-frame basis so events may be more aggregated than
they initially were.

ARCAN_EVENT_SHUTDOWN=keysym:modifiers

This feature provides the ability to specify a panic key combination
that enqueues an exit event rather than the keyboard input. This
allows us to recover and shutdown in controlled environments (egl-dri
platform) when the running application does not provide us with the
means to do so.

Note that this is still subject to the normal event loops, it cannot
help with cases where the active script is stuck in a live-lock
(we'll come to that layer) or the malfunction is in the engine
(hopefully less likely).