(core+) improved multi-display support
The first round of real test cases passed, one possible
issue that we associate with nouveau at the moment is a
tear-like effect but more consistent -- something seems
off timing-wise but what layer being responsible is more
difficult to find out, need to widen the test set.

Anyhow, tested mapping display for a single object,
for world-id and for a rendertarget, all working correctly.

Accelerated cursor- path will likely still be broken,
it shouldn't be a difficult fix -- but we need to update
AWB to handle multiple screens which is the heavier task
by far.