(win32) slowly getting back to a working state
The windows port has been neglected the last
few months and with all the refactoring going on
a lot of things unsurprisingly didn't build or
work at all. This round of commits brings us
back to the point where most normal arcan
features are running, but hijacking and non-auth
connections are still unimplemented and frameservers
are not updated / fixed.