(core+) ripping off one of the many event bandaids
The event management is one of the oldest engine
parts by far and its structure is somewhere between
awful and terrible. This commit is not a fix to that
but a rather intrusive cleanup as part in a larger
effort that will take quite some time to cover.

The bigger side to this, slated around 0.8 or 0.9
is a clean separation in the shmif where only the
events that need to propagate parent<->client are
left, with a portable and serializable communication
path that doesn't have to interleave on the shared
memory page.

This commit will break things for other projects
that use the shmif, though fixes will be quick
in the sense of switching ev.kind -> ev.tgt.kind or
ev.ext.kind depending on the event type, along with
removing the .data union specifier.