(shmif+) necessary refactor
unfortunately breaking things again, some very
odd portability corner cases required that we
move further with the refactor towards events
over socket.

Changed the semantics for descriptor passing,
the operation is now opaque to the user, the
descriptors are part of ioev in the relevant
events, but the user is expected to dup as the
next call to shmif will close.

This will break shmif further on windows,
but since that port will get an overhaul soon
anyhow it is less of an issue.

FossilOrigin-Name: 639e8973a56795d20021fa89a359270585f3961036585cd097540dc5aea7af9a