(core) changed frameset behavior for single-frame
Previously, there were no semantics for dropping a frameset once allocated
and associated with an object. This has rarely caused complications as
it has been possible to create a new object and transfer desired properties,
but this does not work for more complex (3d models, frameservers) situations.

This commit adds the behavior for the specific case of requesting a
frameset of 1 (meaning that the object should only reference its primary
storage) means dropping the currently existing frameset.