(core+shmif) honor initial segment size
This fixes a subtle issue breaking recording frameservers.
Previously, all primary segments where allocated to a fixed
compile-time size and the custom was that the connected
frameserver requested a resize.

When subsegment support was added, the option to hint
desired dimensions followed in order to reduce the number
of resize operations (due to the added connection cost).

The way the shared memory pages were allocated was not updated
to reflect this, which was not noticable on subsegments that
were less than the start size (which was already in the 2M range)
and not on primary segments that started by doing a resize.

For recording frameservers, however, where the resize option
is disabled this would create a crash in, typically, the frameserver.

The solution now is that the hinted dimensions are used when
allocating the shmpage, and when the segment is acquired, the
segment_size is checked against the expected size and remapped
if that deviates. This could possibly also be re-used to drop the
need for overcommit on OSX.