(core) defer higher-level resize to first frame
Previously, as soon as a frameserver got a resize
request approved, the corresponding event was sent
and the video backing store populated with an
empty frame. This have lead to an initial "black flicker"
on highly visible resizes. With this commit, the
resize event and backing store resize is connected to
the push_buffer step (thus when a frameserver has
delivered a frame).

FossilOrigin-Name: 8a2b0245c99dd6263d0c40bf9f817ac171ec89c3318ac0703fe5694a913a85ed