(core+) reworked target_alloc
This is a complex commit with a drawback or two, and requires a Lua API version
bump. There was a known race-condition issue with target_alloc for
external connections as the socket was unlinked at the first connection,
meaning that pending connections could be dropped. Part of the problem
is that we should support rate- and connection limiting, while re-using
as much code paths as possible.

A connection now emits a 'connected' event that also contains the key
and the listening descriptor. If target-alloc is repeated within the
scope of the Lua callback, the socket connection point will be re-used
and any pending connections will continue to be processed.

function cbh(source, status)
	if (status.kind == "connected") then
		con = target_alloc(status.key, cbh);
	end
end);

con = target_alloc("extkey", cbh);

if target_alloc is not called, the connection point will be unlinked and
the descriptor closed, when the cbh goes out of scope. This is similar
to how we handle subsegment requests.

FossilOrigin-Name: 6d92a42c6943cd4c59e715e5606335a0c338523da56833d82490b6a9d8e40a58