(refactor) reorganising frameserver namespace
Step one, dropping the arcan_frameserver_ namespace
in favor of afsrv_, creating subdirectories for each
archetype.

Step two will be fixing the CMake lists to account
for this, and moving / modifying platform/.fsrv list
in here to select for variants.

Step three, compilation cleanups.

This is a setup for then moving dependencies and external
out of the src tree and changing cmake modules to check
external folder for overrides before accepting os-built
in.