(core) event- timing and record/replay fixes
Three bugs are corrected here: one is that platform_event
propagation should be performed even outside logical ticks
to prevent pile-ups (responsible for jerky mouse behavior),

the second (courtesy of johan) is that duplicate ticks were
fired and in the wrong sequence.

and the last one is that serialization of mouse events for
event recording/replay had the wrong index.