(build) build/download static dependencies
For some environments where packaging, build-system etc.
is unrealiable, we may want to download and build the key
arcan dependencies statically. This commit is the first
of two that achieves this:

 1. Re-adding sqlite amalgamation
 2. clone script that pulls down the appropriate gits
    into a preset path
 3. build- options that look at the preset path and uses
    ExternalProject_add instead of FindPackage/FindModule

What is missing is that the LWA audio/openal patches no
long apply and need to be fixed to account for changes to
ashmif, and that LuaJIT/Freetype externalproject conf.
is incomplete.

FossilOrigin-Name: d3e062157c9db9177d181c5f2491df7e643c90836327f7e7182a63cdddade4f0