(shmif) added VIEWPORT, changed REGISTER
This is not implemented in the back-end yet but added the option to
limit the currently active viewport (or defined additional ones) for a
segment. This is to permit multiple views on the same segment (cut down
on transfer costs for subwindows) and to hint other drawing properties
e.g. borders.

Slight tuning to semantics for REGISTER events, will now only ever be
permitted once. Subsequent events will be silently dropped.