(shmif) added convenience pitch field in cont
Note the point that pitch/stride is used rather inconsequentially
graphics programming. The use adopted here is to have the stride for the
byte distance between rows and pitch for the pixel distance. This is
important here as using the stride to index vidp would overshoot unless
cast to a byte- sized format.