(core) subsegment cascade death fixes
Important fix, changes the default state for vobjs to something that
doesn't require a ffunc invoke (debugging leftover) and removes a nasty
UAF that is based on a misinterpretation of cascading effects. It turned
out that subsegments did not check the state of their parent during the
regular poll/control state. This meant that in the rare cases where the
parent died and the subsegment was treated as alive, we would get an
audio refresh (though the video refresh is safe even on a dead page)
that would resolve to a non-existing object.

FossilOrigin-Name: 70f99ab7266199c8688fb736f98e8e6b8804e28aa694e406eef041ea071e3970