(shmif) internal _EXIT handling
Previously, the _EXIT event was only forwarded to the caller. This
commit adds an internal state tracking that does not trigger any
guard_thread kill but prevents further events etc. from being processed.
This also involved a cleanup in semantics for _wait and _poll that is
important, _wait => !0 on success, _poll => < 0 fail, 0 no ev. 1 >
success.