(core+) initial support for uniform groups
Previously, we've been forced to create multiple builds of the same
shader when we want to vary the uniforms active for a shader that is
attached to many objects but with different properties. This push,
though incomplete, adds most of the legwork necessary to support uniform
groups that are disguised as new shader ids, but re-use underlying
shader and just copies / remaps uniforms.

We also drop the distinction between PERSIST and NOPERSIST, though both
lua- shader_uniform calls are still supported, uniform values will
always persist (have local storage) between video subsystem restarts.