added support for querying input capabilities
with ongoing work in making the platform/linux layer more stable and
resilient, a need arose for the ability to query if a keyboard or other
device was present in order to determine if an appl is in such a state
that it cannot operate in a meaningful capacity. This push adds a
function to that, along with other improvements to platform/linux/event.

FossilOrigin-Name: 6795deed0e48e42ae2c2cb9eafaef8c8b5b1d2ec191e44d56e8669e296260280