(platform/linux) VT switching working
There still seem to be some issue when SSH-ing in and then trying to
switch VT, as well as when the resolution of the arcan session differs
from the suggested or native resolution of the previous virtual
terminal. Preliminary testing with running X, Weston and Arcan on
different VTs worked out though.