(core+) typo fixes in linux keymap, added resetdsp
Keyboard mapping device- hell continues, there is an odd corner case
where we switch virtual terminals (press:CTRL, press:ALT, press: F1)
events get queued but no corresponding release events as the device
connection and event subsystem is dead. To deal with that, we add a
reset event to the _display handling on the Lua level which hints that
input state is no-longer "reliable".

Apparently, linux evdev translation is tied to the tty, so when we
disable tty echo etc. we also lose the ability to translate keymaps
correctly. This means that we will have to add support scripts for
converting regular linux keymaps to lua scripts that can be used in
conjunction with the regular keysym- etc. system for localization.
Likewise, the translated event structure need to be reworked a bit to
allow passing of valid UTF-8 sequences through all layers.