(core) fixed rt refc. bug
The simple cause was that the default detach- operation in attachobject
when changing ownership still assumed the src was connected to stdoutp
rather than to a new object.

Similar changes were made to push on persisted objects, and there is a
pending question if rtarget destruction reparenting should reparent on
stdout or not.