(egl/dri) temporary changes to synchronization
The current approach to synchronization is quite insufficient in the
egl/dri/kms module still, a problem that becomes apparent when running
timing sensitive (emulator tests) data sources on multiple monitors
where the clocks don't line up (like 59.575Hz source, 59.912Hz monitor1,
59.511Hz monitor2). This patch works around some of the issues by simply
specifying a primary monitor and not updating secondary ones if they are
still in a pending state when it is time for the new update. The
long-term/ reals solution here is to present multiple synchronization
strategies (that can also interact with CPU governors etc.) to cover the
different cases that gaming, mobile and desktop computing present
respectively. Most interfaces are prepared for that, but still not
enough reliable granuality to do it on a display by display basis.