(lua+) asynch- state races in sys-collapse
first in multiple commits needed to fix hand-over in sys-collapse that
retains frameserver control. The big problem is that there might be
events in-flight that gets tagged with lua-state tracking callbacks that
then gets destroyed or "use-after-free-but-not-really" and mapped to the
wrong callback later. This patch fixed the immediate problem of crashes
from "calling nil value" but not the problem of popcontext -> EXIT event
in frameserver. We also miss some sematics to TARGET_COMMAND_RESET that
would communicate that some state information (labels, coreopts,
dimensions etc.) are gone and need to be rebuilt.