(platform) support for notifying about recovery
crash recovery gave undefined video results on egl-dri platform and
others in multiscreen configuration as displays were mapped to video
objects that were popped off-screen and script-related state trackers
was unaware of the state of previously mapped displays. This patch adds
an explicit call that comes after the adopt- stage which fixes this
issue.