(shmif) return _EXIT event even on *ks
Previous behavior has been for all functions to return failstate on
killswitch, which may lead to a race depending on the shutdown order ein
the parent. To prevent a possible descriptor leak (from pending- state
in context) and to allow the client to shutdown properly, allow _poll
and wait to return correctly for _EXIT in the event of killswitch.