(core+) added autoclock implementation
Default is now that most autoclock events that are periodic are handled
inside the engine and will never be exposed to the script event handler.
This behavior is a toggle flag for the frameserver, but there are few
cases where it is useful to turn it off.

FossilOrigin-Name: c94c5c328996ce173f2fa3709ad0da31e54c899ef4f6a88a7c202015f75b4a48