(lua) target_alloc, allow subseg- type enforcement
Most of the time, it makes little sense for a force-pushed subsegment to
have a specific time, if a connected frameserver wants/can handle a
specific type it should request it, or it will be of a special (record,
clipboard) type that goes through other functions.

This patch adds support for specifying the debug or accesibility
subtype when pushing a segment as that is not really to be treated
as feature-reaction but rather event-data relationship.

FossilOrigin-Name: ed5565208239b5c7d3623008502f58cf9f4439a2f3de30b69835fe9f3de96956