(tools) added first draft of conv- tool
For egl-dri with linux platform, we need some table format for use with keysym.lua to handle
internationalization and keyboard layouts. This is the first such attempt that uses loadkeys
to create a c file that we include and convert into a .lua file (offline process, only need
to do once). Currently untested and don't handle advanced (fun->str mapping and diactretics)