(core+) drop clones, refactor vint, improve fset
This patch first drops the internal use and support for clones
(instance_image), it was too complicated and underused to be worth the
cost. The legacy suppory script gets a backwards compatible workaround.

Framesets now retain their own set of texture coordinates that are
either the normal default or a copy of the ones active on the vobj at
the time setasframe was called. This allows for some cases like
round-robin packing into the same texture and animated spritesheet using
the same backing store.