(core+) lessen db restrictions, expose hash func
The database key was purposefully restricted to a very limited character
set, to be able to be accessed from the keyboard regardless of state of
encoding etc. This push lessens that slightly to allow for the set
needed for base64 encoded key values.

In addition, we started exposing a hash utiity function in the util:
namespace (lua) that currently only does a simple djb hash but will be
extended in 0.6 to cover the more relevant ones (SHA-1,3)