(core) harder enforcement of playstate
Since dropping the framequeue, suspend/resume has always worked
like a hint, the tracking of PLAYSTATE between normal and playing
have typically been ignored.

This also required a change to the spin-lock way of _resize as
there is a small race that would lead to high consumption in a
suspendede client.