(lwa) VRES_AUTOFONT for FONTHINT
like with displayhint, we need to propagate some information for
appls to be able to handle that the default font behavior has changed
so that interface recalculations can be done. We do this similar to
VRES_AUTORES in the sense of a global that is called if set.

FossilOrigin-Name: bd13d5814f6f6590df56469d10fa0fc231f35d903a2fe5bd60e8f7b92b0cb9e1