(core) default database behavior changed to ~/.arcan/arcan.sqlite
Previous behavior has been to create the database in the respath,
which is 'suboptimal' (dangerous and insecure as well) and has
mainly been ignore since most cases have been with an explicit
dbpath. To make this a bit friendlier, the platform layer now
presents a default safe dbpath which is used as second fallback.

For POSIX, this defaults to getenv("HOME")/.arcan/arcan.sqlite.