(core+shmif) dirty-upd + bufferswap preparations
Next in line for terminal and shmif updates requires a bump in
shmif versioning. We add the option to synch only a subregion,
which should help in some remoting clients and for the terminal.

Then we continue support for splitting vbuf/abuf into multiple
on resize, so it is possible to semi-transparently alternate
region that is being synched and transfers to reduce the cascade
from shmif_signal blocking. This will improve the situation for
latency sensitive buffering in game frameserver.

FossilOrigin-Name: 4855d894a23b24b1f119f2fab9d009785268334eaa88c5e147028ae263e350b6