(shmif++) first push on resize-synch work
There are several problems with the current approach
to resize behavior in the shared memory inteface:

1. 1V/1A bufffer prevents other strategies
It would be better to be able to configure the amount of audio and video
buffers used so that signal can return immediately and the client can
produce the next frame itself without running into the vsynch-block or
resorting to own buffer management with extra copy. Some cases don't
call for any audio buffering, and shouldn't have to allocate one.

2. Handle passing still requires an equally sized backing store
It should be possible to set segment output dimensions without having
a corresponding segment backing store. This cuts down on memory use
when we have accelerated passing.

This push is the first of several that attempts to fix this,
first by changing the way offsets and buffer sizes are calculated.

Secondly by adding 'apending/vpending' fields to indicate desired
buffer counters and as a mask of populated buffers.

Thirdly by having a loser fall path for _resize and updating _resize
calls with an extended version that permits specifying buffer sizes.

Lastly with a flag to signal calls as to how the buffers should be
treated.