(core+build) evicted WIN32
This has been pending for a long time. If it was ever in doubt that maintaining
a windows port was worth it, 10+ and other changes pushed the tipping
point. Until there is a 'sane' build environment (LLVM/CLANG)on windows,
and a supporting POSIX implementation (support Midipix!) that is complete enough
to support the project, the windows port is stashed in the legacy-junk pile where
it belongs. Begone #ifdef WIN32 shit.