(core) fixed frameset regression
An older rework to skip of the draw surface was entirely outside visible
view inadvertently introduced a re-order for the stage when a vobj-
shader was active in relation to its vstore. Normally this didn't have
an effect but for multitexture- store, it meant that the uniforms
corresponding to setting the relevant TUs resolved to the wrong shader.

FossilOrigin-Name: fa44812c16133f777dce18a12c5fc5ae44e1945da99712492ff4c2a536a81663