(platform/linux) inotify race workaround
Among the (many) problems with the evdev input approach is that
the inotify on monitor folder can be subject to a subtle race if
the mknod() caller doesn't have the right euid etc. set for the
permissions to be correct immediately. Add a small retry queue
to account for such cases.

Also switched 'rescan/collision' check to use path instead of
node ID hash as there's too high a chance for collision for some
devices (K65 keyboard has been very vulnerable to this).