(lua) require _adopt to explicitly return true
Previously, _adopt function would just assume to be successful.
This turned out problematic in the cases where you need to delete some
vid that should not be adopted. Since it's not safe to call delete_image
on the vid during adopt, this should just return false/nil and it
will be added to the kill-queue.