(arcterm) more work on terminal responsiveness & tradeoffs
This is the gift that will never stop giving. Whack one mole and about
ten others take its place. It's not that the ambition is to ever pass
all of vttest (because life is too short), but the current bag of fun
that is state-machine v. rendering race conditions that seem to mostly
manifest with using vim has all the traits necessary for the 'yes, I'll
debug you later, never to return' escape plan.

So yes, I'll debug you later.