(arcterm) nottf build fixes, italic state support
The problems with disappearing characters etc. in vim persists and seems
hard to reproduce outside of it. To not waste any more time on the
issue, consider a switch to a different state machine (libvterm seem
more actively maintained). At the same time, other programs don't seem
to have much issue with the current state of affairs, and neither does
nvim.