(engine+lua) next round of font improvements
This is the first in a new round of font rendering changes. Many
of the internal interfaces for specifying hinting, font and so on
are extended with the ability to specify a fallback fonti - intended
for emojis and other exotic cases where the main font may not cover
the glyphs needed.

Inside the SDL_ttf derived arcan_ttf, we have now stripped away all
UCS2 support in favor of UTF8 at borders and UCS4 internally, and
subsequently removed all render functions that dealt with the old
abstractions.

This does not fully fix the quality issues with hinting, sizes and
glyph coverage, but it is a start.

FossilOrigin-Name: 0a58b57d9b91460e58796c07efe6c7a841aeecb64f05929ee195ca70ea1925a8