(core/build) font fallback improvements
Build-system changed for static freetype build to indicate whether or
not the libpng dependency is wanted or not (needed for bitmapped
glyphs). The internal interfaces now operator on the fallback- chain
approach where additional fonts can be used when the default one is
missing a particular glyph. Roughly speaking, we are still missing:

1. scaling (fallback and regular) bitmap to match primary em size.
2. consitency checking hinting changes
3. adding a direct U32 glyph drawing to speed up arcterm glyph
processing.