(core+arcterm) split string and glyph rendering
Previously, all rendering went through the UTF-8 to UCS4 etc. conversion
and sizing, even when not needed. This has been fixed by moving the
rendering to a UNICODE glyph call (it is still ugly but that's lower on
the list).

Arcterm now supports fallback font and associated glyph rendering,
albeit that the sizes are still not entirely correct. This also means that
the ever elusive midnight commander borders are drawn in UTF-8 mode.