(core) updated frameserver_enter guard
This is not for the faint of heart and is a rather horrible approach,
but the critical- section approach to SIGBUS-on-mmap+ftruncate
protection needed had siglongjmp set a context with the wrong contents.
In this patch, the interface is changed to allow a jmp_buf context
chain, so that siglongjmp can jump to the cleanup reaction, and then
chain to the proper error-code return.