(core) updated frameserver_enter guard
This is not for the faint of heart and is a rather horrible approach,
but the critical- section approach to SIGBUS-on-mmap+ftruncate
protection needed had siglongjmp set a context with the wrong contents.
In this patch, the interface is changed to allow a jmp_buf context
chain, so that siglongjmp can jump to the cleanup reaction, and then
chain to the proper error-code return.

FossilOrigin-Name: 042815d0b24c3f0044e98feb4131278f092ca13c3717a8216e0a0d717e6761d6