(arcterm) changed resize heuristic
This was the source of one subtle yet annoying drawing bug that made the
text in terminals for durden look weird. Previously, to prevent the
onslaught of resize events during drag resize, there was a rule that
only issued a resize if the changed dimensions were GE to single cell
dimensions. When we started communicating desired resize step, this rule
was superflous but remained in effect. So when autolayout resizing were
less than cell dimensions, we'd suffer precision errors with the crop
shader used.