(egl-dri) continue card-scan when no active disp
If ARCAN_VIDEO_DEVICE is not set, the default card search behavior has
simply been to sweep /dev/dri/card* and using the first one that could
be accessed. If this card lacked connected displays, we would fail and
exit. This change resumes the card scan at the next index in order to
work in cases where there are multiple cards but only one (not index 0)
with connected displays.