(lua) TOCrTOU fix for open_nonblock in fifo-write
Two bugfixes in open_nonblock() write-only for FIFOs. One is that
if a fifo already exists in the matching path, it should still be
opened. The other is that failure- during write should fall back
to try and reconnect with the same FIFO path and accept open/write
even without a connected consumer.