(core/lua) big rendertarget- update changes
This is a prelude towards less awful dirty- region and controlled
updates for rendertargets with different clocks. _forceupdate calls
can now be used to change the refresh-clock for a rendertarget, and
it is now possible to enumerate vids attached to a rendertarget.
set_context_attachment behavior is changed to support returning a
reference to the current primary attachment.