(build) FreeBSD console/mouse driver preparations
After a weekend of testing, it seems that the KMS support in FBSD is good
enough to start adding support for a local / terminal input platform to
match. This push renames the 'linux' input platform to the more fitting 'evdev'
(as it also works on DragonFlyBSD) and adds a stub for FBSD console.

FossilOrigin-Name: 7e52f16664f436c8c039f9af16624cb92ef53b7920dc15131845bfec597bd8e3