(platform/evdev) proper handling of SIGTERM
SIGTERM was just set to the default behavior, which did not work well
on the egl-dri platform, leaving the VT in an unusable state. This patch
adds SIGTERM to the TTY-switch internal FD and maps it to a normal
SHUTDOWN event. This gives the event handler in the appl time to shut
down, and the graphics state can be cleaned up gracefully.

FossilOrigin-Name: ec0b9e6cd9c7b06f7650ee536678beded8c0da851294ba2c6ed461ed88083e9f