(egl-dri) didn't set correct flag for the current mode
the 'primary' field of the query_displaymodes set of calls should
be used to indicate if this is the active display mode or not. This
was ignored by the egl-dri platform, making it impossible for a
script to know the initial display mode state.

FossilOrigin-Name: e1a6d4f5d81012593bb77023f53b4877a4fde9376b614afe5a7352e2af71a0a8