(shmif) shmif-migration and auto-reconnect
This is a semi- working push for one of the key features needed
for future remoting and resilience. We add support for primary-
segment migration from one connection path to another by first
connecting to the new path, then transfering relevant state.

Then we combine this with the guard-thread so that when the dms or
event-queue ks is pulled, we instead 'migrate' to the same primitive
we used to initiate the connection. This means that if the server
should die/crash, we retain some/most state (but need a little bit
of cooperation from the lib- user) and indicate that something went
south by adding a _RESET event.

There are definitely kinks- and tests to work out before this is
useful, but so far, nothing too major.