(shmif) migrate-on-dms shouldn't trigger on _EXIT
There are a few migration issues remaining, one is partial corruption
of -priv in some hard-to-reproduce edge conditions for terminal. The
others relate to the poor use of dms in frameservers that need to be
refactored. To extend the testset, some _lua/shmif that also relate to
card/render-node migration need to be implemented (basically in-flight
hints on render-node, scanout use or acceleration list).

FossilOrigin-Name: 853a0ce26625df44768b674f47396a4d87417cbfe932bea8e2b8c62c4cf37abc