(shmif) added _ext shmif sub library
This is the first part in breaking out the headless- handle-
passing foreplay to a part of shmif (if it is explicitly requested)
in order to make it easier to manage external processes that wants
to use accelerated handles, e.g. fsrv_game, arcan_lwa.

This is also a part in adding support for vulkan- producers and
dynamiYc migration between GPUs.