(core+) reworking audio synch and sample-rate
This is one of many preparations for a more sane and efficient
synchronization framework. We start by allowing shmif- extended
resize negotiation to change samplerate. By doing so we can start
dropping buffering and resampling from frameservers and move that
to the audio platform. This unfortunately breaks the recording
of auidio sources with deviating sample rates, but that will
be fixed shortly.