(hijack/x11) primitive glxgears showing
This is not that much of a serious attempt to get an 'xlib->
arcan' implementation in a .so, but more of a learning experiment
and looking for features X supports that weren't considered in
shmif. At most it would complement the cases where we get xlib as
a parasitic dependency for something trivial.

This push adds enough 'red tape' to be able to get glxgears
running in a limited fashion (no input, no proper display properties
etc.)

FossilOrigin-Name: 5da79ef7f35f23a3274fcb0f6782e11900728ceb0c4936e1f43c9b5dfa4c636a