(hijack/x11) primitive glxgears showing
This is not that much of a serious attempt to get an 'xlib->
arcan' implementation in a .so, but more of a learning experiment
and looking for features X supports that weren't considered in
shmif. At most it would complement the cases where we get xlib as
a parasitic dependency for something trivial.

This push adds enough 'red tape' to be able to get glxgears
running in a limited fashion (no input, no proper display properties
etc.)