(shmif) added WIDGET SEGID type, OUTPUTHINT
In preparation for rendertarget->subsegment support in LWA, we
need another SEGID type for dockable widgets etc. (both the
calc/weather style and good ol' window-maker docklets fall in
this category).

Also added an OUTPUTHINT that covers capability rather than
status. That is, DISPLAYHINT is used to convey the current
properties, while-as OUTPUTHINT indicates the maximium. Later
patches will wire the corresponding parts to target_displayhint
(new field in the table), to target_alloc and to accept_target.