(lwa) work on arcan_definetarget
next round of improvements to lwa- mode is to be able to request
and allocate new subsegments (icon, cursor, titlebar etc.). That
is a pretty heavy feature and need to be broken down into a number
of steps. This push adds the necessary boilerplate that needs to
be fleshed out.