(core) preparation for event-prequeue
A notable source of startup time and latency is the negotiation
for initial size, font and display properties. To fully combat
this we need to extend the Lua api to provide surface and display
properties immediately, and for the client to run through the
event queue as soon as possible.